testing/web-platform/tests/dom/nodes/Comment-constructor.html

<!doctype html>
<meta charset=utf-8>
<title>Comment constructor</title>
<link rel=help href="http://dom.spec.whatwg.org/#dom-comment">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  var comment = new Comment();
  assert_equals(Object.getPrototypeOf(comment),
                Comment.prototype, "Prototype chain: Comment");
  assert_equals(Object.getPrototypeOf(Object.getPrototypeOf(comment)),
                CharacterData.prototype, "Prototype chain: CharacterData");
  assert_equals(Object.getPrototypeOf(Object.getPrototypeOf(Object.getPrototypeOf(comment))),
                Node.prototype, "Prototype chain: Node");
}, "new Comment(): prototype chain");

test(function() {
  var comment = new Comment();
  assert_true(comment instanceof Node, "Should be a Node");
  assert_true(comment instanceof CharacterData, "Should be a CharacterData");
  assert_true(comment instanceof Comment, "Should be a Comment");
}, "new Comment(): instanceof");

test(function() {
  var comment = new Comment();
  assert_equals(comment.data, "");
  assert_equals(comment.nodeValue, "");
  assert_equals(comment.ownerDocument, document);
}, "new Comment(): no arguments");

var arguments = [
  [undefined, ""],
  [null, "null"],
  [42, "42"],
  ["", ""],
  ["-", "-"],
  ["--", "--"],
  ["-->", "-->"],
  ["<!--", "<!--"],
  ["\u0000", "\u0000"],
  ["\u0000test", "\u0000test"],
  ["&amp;", "&amp;"],
];

arguments.forEach(function(a) {
  var argument = a[0], expected = a[1];
  test(function() {
    var comment = new Comment(argument);
    assert_equals(comment.data, expected);
    assert_equals(comment.nodeValue, expected);
    assert_equals(comment.ownerDocument, document);
  }, "new Comment(): " + format_value(argument));
});

test(function() {
  var called = [];
  var comment = new Comment({
    toString: function() {
      called.push("first");
      return "text";
    }
  }, {
    toString: function() {
      called.push("second");
      assert_unreached("Should not look at the second argument.");
    }
  });
  assert_equals(comment.data, "text");
  assert_equals(comment.nodeValue, "text");
  assert_equals(comment.ownerDocument, document);
  assert_array_equals(called, ["first"]);
}, "new Comment(): two arguments")

async_test("new Comment() should get the correct ownerDocument across globals").step(function() {
  var iframe = document.createElement("iframe");
  iframe.onload = this.step_func_done(function() {
    var comment = new iframe.contentWindow.Comment();
    assert_equals(comment.ownerDocument, iframe.contentDocument);
  });
  document.body.appendChild(iframe);
});
</script>
